% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazoom.amazonia-package.R
\docType{package}
\encoding{UTF-8}
\name{datazoom.amazonia-package}
\alias{datazoom.amazonia}
\alias{datazoom.amazonia-package}
\title{datazoom.amazonia: Simplify Access to Data from the Amazon Region}
\description{
Functions to download and treat data regarding the Brazilian Amazon region from a variety of official sources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.econ.puc-rio.br/datazoom/}
}

}
\author{
\strong{Maintainer}: Laura Tavares Regadas \email{lauratregadas@gmail.com}

Authors:
\itemize{
  \item Gustavo Gonzaga \email{gonzaga@econ.puc-rio.br}
  \item Igor Rigolon Veiga \email{igor.rilave@hotmail.com}
  \item Guilherme Jardim \email{gnjardim1@gmail.com}
  \item Daniel AC Barbosa \email{daniel.adrianocb@gmail.com}
  \item Bruno Alcantara Duarte \email{brunoalcanduarte@gmail.com}
  \item Fredie Didier \email{fdidier@terra.com.br}
  \item Tito Bruni
  \item Luiz Guilherme Lopes Moussatche \email{luizguilherme.lopesm@gmail.com}
  \item Victor Aliende da Matta \email{victor.al.matta@gmail.com}
  \item Anna Carolina Dutra Saraiva \email{annadutrasaraiva@gmail.com}
  \item Arthur Carvalho Brito \email{arthurcbps@gmail.com}
  \item Francisco de Lima Cavalcanti \email{francisco.lima.cavalcanti@gmail.com}
  \item Maria Mittelbach \email{mariamittelbach@gmail.com}
  \item Felipe de Moraes Nunes \email{felipedmoraesn@gmail.com}
  \item Joao Paulo Santos P. Barbosa \email{joao.31582129@gmail.com}
  \item Bernardo Furlanetto Sieira \email{bernardo.sieira@gmail.com}
}

Other contributors:
\itemize{
  \item DataZoom (PUC-Rio) \email{datazoom@econ.puc-rio.br} [funder]
}

}
\keyword{internal}
