% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_biomas_cobertura_va.R
\encoding{UTF-8}
\name{datazoom_cobertura_mapbiomas_va}
\alias{datazoom_cobertura_mapbiomas_va}
\title{datazoom_cobertura_mapbiomas_va}
\usage{
datazoom_cobertura_mapbiomas_va(
  aggregation_level = c("municipality", "state", "municipio", "estado"),
  path = NULL,
  code_state = NULL,
  code_mun = NULL,
  covering = NULL,
  type = c("stacked", "normal", "empilhado"),
  year_begin = NULL,
  year_end = NULL
)
}
\arguments{
\item{aggregation_level}{A string that indicates the level of aggregation of the data. It can be by municipality or state}

\item{path}{A string indicating where the raw data is in your computer. The default is NULL which means the data will be extracted directly from the website}

\item{code_state}{Output contains only data from the state selected
Input has to be IBGE's numeric coding for the state desired}

\item{code_mun}{Output contains only data from the municipality selected (not available for aggregation_level=='state')
Input has to be IBGE's numeric coding for the municipality desired}

\item{covering}{Output contains only data over the selected covering
Input has to be the code of the desired covering}

\item{type}{Decide if the output should have a column for each year (normal) or a single column for the areas of all years (stacked | empilhado)}

\item{year_begin}{A numeric object containing the first year desired in the data base}

\item{year_end}{A numeric object containing the last year desired in the data base}
}
\value{
A data base with data containing the area of each selected type of soil covering in each selected year
}
\description{
Download and filter data on type of soil covering by year
}
\examples{
datazoom_cobertura_mapbiomas_va(aggregation_level = 'municipality', path = '/Desktop',
                             code_state = 11, code_mun = NULL, covering = 3,
                             type = 'stacked', year_begin = 2000, year_end = 2010)
}
\author{
DataZoom, Department of Economics, Pontifical Catholic University of Rio de Janeiro
}
