% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazoom_mapbiomas_transicao.R
\encoding{UTF-8}
\name{datazoom_transicao_mapbiomas}
\alias{datazoom_transicao_mapbiomas}
\title{datazoom_transicao_mapbiomas}
\usage{
datazoom_transicao_mapbiomas(
  path = NULL,
  code_state = NULL,
  covering_from = NULL,
  covering_to = NULL,
  type = c("stacked", "normal", "empilhado"),
  year_diff = NULL
)
}
\arguments{
\item{path}{A string indicating where the raw data is in your computer. The default is NULL which means the data will be extracted directly from the website}

\item{code_state}{Output contains only data from the state selected
Input has to be IBGE's coding for the state desired}

\item{covering_from}{Output contains only data of the transition from the covering selected to all other coverings
Input has to be the code of the desired covering}

\item{covering_to}{Output contains only data of the transition from all coverings to the covering selected
Input has to be the code of the desired covering}

\item{type}{Decide if the output should have a column for each year (normal) or a single column for the areas of all years (stacked | empilhado)}

\item{year_diff}{A numeric object containing the year differential desired in the data base, or have all included (NULL)}
}
\value{
A data base with data containing the area of each selected type of soil covering transition in each selected year
}
\description{
Download and filter data on transition of types of soil covering by year
}
\examples{
datazoom_transicao_mapbiomas(path = NULL,
                             code_state = "PA", covering_from = 3, covering_to = 19,
                             type = 'normal', year_diff = 5)
}
\author{
DataZoom, Department of Economics, Pontifical Catholic University of Rio de Janeiro
}
