% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Title}
\usage{
download(
  data = NULL,
  source = "ibge",
  geo_id = "municipality",
  time_id = 2017:2018
)
}
\arguments{
\item{time_id}{}
}
\value{
A tibble
}
\description{
Title
}
