% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fines_ibama.R
\encoding{UTF-8}
\name{fines_data_agg}
\alias{fines_data_agg}
\title{Compiles environmental fines data in the Amazon Region from IBAMA}
\usage{
fines_data_agg(
  time_unit = c(.data$year, .data$month),
  geographic_unit = .data$city,
  years = 2005:2020,
  language = "english"
)
}
\arguments{
\item{time_unit}{Temporal level of aggregation, choose "year" and/or "month".
If \code{language = 'portuguese'}, pick "ano" and/or "mes'"}

\item{geographic_unit}{Geographical level of aggregation, choose "city" and/or
"state". For portuguese version, "municipio" and/or "uf"}

\item{years}{Time period to be considered. Contemplates 2005 to 2020.}

\item{language}{Language used in returned dataset. Use \code{language = "portuguese"} or
\code{language = "english"}}
}
\value{
A dataframe with counts for variables related environmental fines and
infractions by \code{time_unit}-\code{geographic_unit} chosen. Data include
counts for total number of infractions, infractions that already went to trial, and
number of unique perpetrators of infractions on the given place-time period.
}
\description{
Original data comes from \href{https://servicos.ibama.gov.br/ctf/publico/areasembargadas/ConsultaPublicaAreasEmbargadas.php}{IBAMA's website}
}
\examples{

fines_data_agg(
  time_unit = year,
  geographic_unit = state, years = 2005:2006
)

fines_data_agg(
  time_unit = c(year, month),
  geographic_unit = c(city, state)
)

fines_data_agg(
  time_unit = ano,
  geographic_unit = municipio, years = c(2010, 2012),
  language = "portuguese"
)
}
