% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fines_ibama.R
\encoding{UTF-8}
\name{fines_ibama}
\alias{fines_ibama}
\title{Compiles environmental fines data in the Amazon Region from IBAMA}
\usage{
fines_ibama(
  download_data = TRUE,
  download_directory,
  load_from_where,
  time_unit = c("year", "month"),
  geographic_unit = "city",
  years = 2005:2021,
  language = "english"
)
}
\arguments{
\item{download_data}{If \code{TRUE}, downloads the data from the original source.
Else, loads data from \code{load_from_where} directory.}

\item{download_directory}{Directory where raw data should be stored.
Used in case \code{download_data = TRUE}.}

\item{load_from_where}{Original that should be loaded. Used in
case \code{download_data = FALSE}.}

\item{time_unit}{Temporal level of aggregation, choose \code{year} and/or \code{month}.
If \code{language = 'portuguese'}, pick "ano" and/or "mes'"}

\item{geographic_unit}{Geographical level of aggregation, choose \code{city} and/or
\code{state}. For Portuguese version, \code{municipio} and/or \code{uf}}

\item{years}{Time period to be considered. Contemplates 2005 to 2020.}

\item{language}{Language used in returned dataset. Use \code{language = "portuguese"} or
\code{language = "english"}}
}
\value{
A data frame with counts for variables related to environmental fines and
infractions by \code{time_unit}-\code{geographic_unit} chosen on the Amazon region.
Data include
counts for total number of infractions, infractions that already went to trial, and
number of unique perpetrators of infractions on the given place-time period.
}
\description{
Original data comes from \href{https://servicos.ibama.gov.br/ctf/publico/areasembargadas/ConsultaPublicaAreasEmbargadas.php}{IBAMA's website}
}
