% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{initialize_ee}
\alias{initialize_ee}
\title{Load Interface for Google EE}
\usage{
initialize_ee()
}
\value{

}
\description{
Execute the ee_Initialize command. You will be directed to your Google Earth
Engine and get access to it using rgee package
}
