% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipios.R
\name{load_amazon_employment}
\alias{load_amazon_employment}
\title{Gets formal employment data for the Brazilian Legal Amazon}
\usage{
load_amazon_employment(
  years,
  space_aggregation = "municipality",
  language = "eng"
)
}
\arguments{
\item{years}{A numeric vector with years of interest. Supported years 2006-2018.}

\item{space_aggregation}{A string that indicates the level of aggregation of the data. It can be by "Municipality" or
"State"}

\item{language}{A string that indicates in which language the data will be returned. The default is "eng", so your data will be returned in English.
The other option is "pt" for Portuguese.}
}
\value{
A \code{tibble} with (total) Salary (in R$), (number of) Employed people and (number of) Firms.
}
\description{
Gets formal employment data for the Brazilian Legal Amazon
}
\examples{
\dontrun{
load_amazon_employment(2017)
}
}
