% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipios.R
\name{load_amazon_income}
\alias{load_amazon_income}
\title{Gets income data for the Brazilian Legal Amazon}
\usage{
load_amazon_income(years, space_aggregation = "municipality", language = "eng")
}
\arguments{
\item{years}{A numeric vector with years of interest. Supported years are 2000, 2010 (census years).}

\item{space_aggregation}{A string that indicates the level of aggregation of the data. It can be by "Municipality" or
"State"}

\item{language}{A string that indicates in which language the data will be returned. The default is "eng", so your data will be returned in English.
The other option is "pt" for Portuguese.}
}
\value{
A \code{tibble} with Employed population, Mean_Income and Median_Income (in R$).
}
\description{
Gets income data for the Brazilian Legal Amazon
}
\examples{
\dontrun{
load_amazon_income(2010)
}
}
