% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ben.R
\encoding{UTF-8}
\name{load_ben}
\alias{load_ben}
\title{BEN}
\usage{
load_ben(dataset, raw_data = FALSE, time_period = 2004:2021, language = "pt")
}
\arguments{
\item{dataset}{A dataset name ("HS92").}

\item{raw_data}{A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.}

\item{time_period}{A \code{numeric} indicating for which years the data will be loaded, in the format YYYY. Can be any vector of numbers, such as 2010:2012.}

\item{language}{A \code{string} that indicates in which language the data will be returned. Portuguese ("pt") and English ("eng") are supported.}
}
\description{
National Energy Balance
}
\examples{
\dontrun{
# download treated data for 2016
clean_ben <- load_ben(
  dataset = "EPE",
  raw_data = FALSE,
  time_period = 2016
)
}

}
