% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cempre.R
\name{load_cempre}
\alias{load_cempre}
\title{CEMPRE - Central Register of Companies}
\usage{
load_cempre(
  dataset = "cempre",
  raw_data = FALSE,
  geo_level = "municipality",
  time_period = 2017:2018,
  language = "eng",
  sectors = FALSE,
  legal_amazon_only = FALSE
)
}
\description{
Loads information on companies and other organizations and their respective formally constituted local units, registered with the CNPJ - National Register of Legal Entities. Data is available from 2006 to 2019. See \url {https://sidra.ibge.gov.br/pesquisa/cempre/tabelas}

@encoding UTF-8

@param dataset A dataset name ("cempre").
@param raw_data A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.
@param geo_level A \code{string} that defines the geographic level of the data. Defaults to national level, but can be one of "country", "state" or "municipality". See documentation of \code{sidrar}.
@param time_period A \code{numeric} indicating what years will the data be loaded in the format YYYY. Can be a sequence of numbers such as 2010:2012.
@param language A \code{string} that indicates in which language the data will be returned. Currently, only Portuguese ("pt") and English ("en") are supported. Defaults to "en".
@param sectors A \code{boolean} that defines if the data will be return separated by sectors (\code{TRUE}) or not (\code{FALSE})
@param legal_amazon_only A \code{boolean} setting the return of Legal Amazon Data (\code{TRUE}) or Country's Data (\code{FALSE})

@return A \code{tibble} with a panel of N x T observations by municipality-year
}
\examples{
\dontrun{
# download treated data from 2006 to 2019
cempre_all <- load_cempre(dataset = "cempre", raw_data = FALSE, time_period = 2006:2019)

# download raw data from 2006 to 2019
raw_cempre_all <- load_cempre(dataset = "cempre", raw_data = TRUE, time_period = 2006:2019)
}

}
