% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{load_climate}
\alias{load_climate}
\title{TerraClimate}
\usage{
load_climate(
  dataset,
  raw_data = FALSE,
  time_period,
  language = "eng",
  legal_amazon_only = FALSE
)
}
\arguments{
\item{dataset}{A dataset name}

\item{raw_data}{A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.}

\item{time_period}{A \code{numeric} or \code{date} indicating what years will the data be loaded in the format YYYY. Can be a sequence of numbers such as 2010:2012.}

\item{language}{A \code{string} that indicates in which language the data will be returned. Currently, only Portuguese ("pt") and English ("eng") are supported. Defaults to "eng".}

\item{legal_amazon_only}{A \code{boolean} setting the return of Legal Amazon Data (\code{TRUE}) or Country's Data (\code{FALSE}). Defaults to \code{FALSE}}
}
\value{
A \code{tibble}
}
\description{
TerraClimate
}
\examples{
\dontrun{
load_climate()
}

}
