% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comex.R
\encoding{UTF-8}
\name{load_comex}
\alias{load_comex}
\title{Loads foreign trade data from the Brazilian government from 1986 to the present}
\usage{
load_comex(
  year,
  ncm = FALSE,
  sh4 = FALSE,
  exp = FALSE,
  imp = FALSE,
  language = "pt"
)
}
\arguments{
\item{year}{A \code{vector} indicating what years will the data be loaded}

\item{ncm}{A \code{boolean} that when set to \code{TRUE} returns data frame from NCM system, default is \code{FALSE}}

\item{sh4}{A \code{boolean} that when set to \code{TRUE} returns data frame from SH4 system, default is \code{FALSE}}

\item{exp}{A \code{boolean} that when set to \code{TRUE} returns export data, default is \code{FALSE}}

\item{imp}{A \code{boolean} that when set to \code{TRUE} returns import data, default is \code{FALSE}}

\item{language}{A \code{string} that indicates in which language the data will be returned. The default is "pt", so your data will be returned in Portuguese. Currently, only Portuguese and English are supported.}
}
\value{
A \code{list} of data frames
}
\description{
Loads foreign trade data from the Brazilian government from 1986 to the present
}
\examples{
\dontrun{
load_comex(c(2000, 2001, 2002), 
  ncm = TRUE,
  exp = TRUE)
load_comex(c(1997, 2004, 2008),
  sh4 = TRUE,
  imp = TRUE,
  language = "eng")
}

}
\author{
DataZoom, Department of Economics, Pontifical Catholic University of Rio de Janeiro
}
