% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasus.R
\name{load_datasus}
\alias{load_datasus}
\title{DATASUS}
\usage{
load_datasus(
  dataset,
  time_period,
  states = "all",
  level = "municipality",
  raw_data = FALSE,
  language = "eng"
)
}
\arguments{
\item{dataset}{A dataset name, can be one of ("datasus_sim", "satasus_sih", "datasus_cnes"), or subsets thereof. For more details, try \code{vignette("DATASUS")}.}

\item{time_period}{A \code{numeric} indicating for which years the data will be loaded, in the format YYYY. Can be any vector of numbers, such as 2010:2012.}

\item{states}{A \code{string} specifying for which states to download the data. It is "all" by default, but can be a single state such as "AC" or any vector such as c("AC", "AM").}

\item{level}{A \code{string} choosing whether to aggregate the data by "municipality" or to keep all the "individual" variables.}

\item{raw_data}{A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.}

\item{language}{A \code{string} that indicates in which language the data will be returned. Currently, only Portuguese ("pt") and English ("eng") are supported. Defaults to "eng".}
}
\value{
A \code{tibble}
}
\description{
DATASUS
}
\examples{
\dontrun{
mortality_rj <- load_datasus(dataset = "datasus_sim", time_period = 2010, states = "RJ")
}

}
