% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrad.R
\encoding{UTF-8}
\name{load_degrad_raw}
\alias{load_degrad_raw}
\title{Loads degradation data from INPE.}
\usage{
load_degrad_raw(source)
}
\arguments{
\item{source}{A number of different sources are supported:

Passing a numeric \code{vector} of years will download the corresponding data from the INPE website. Available years are 2007-2016.

Passing a \code{string} with a directory's path will read data from all shapefiles in the directory.

Alternatively, \code{source} may be a list of full file paths, or anything else readable by \code{sf::read_sf()}.}
}
\value{
A list of \code{tibble}.
}
\description{
Loads degradation data from INPE.
}
\examples{
load_degrad_raw(2016)

load_degrad_raw(c(2013, 2014, 2016))

\dontrun{
load_degrad_raw("~/Downloads")

load_degrad_raw("~/Downloads/degrad2016_final_shp/DEGRAD_2016_pol.shp")
}
}
\seealso{
\link{load_degrad} for loading and treating the data.
}
\author{
DataZoom, Department of Economics, Pontifical Catholic University of Rio de Janeiro.
}
