% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deter.R
\encoding{UTF-8}
\name{load_deter_raw}
\alias{load_deter_raw}
\title{Loads INPE deforestation data.}
\usage{
load_deter_raw(source, years = "all")
}
\arguments{
\item{source}{A string indicating where the data will be drawn from.

It can be "Amazonia" or "Cerrado", in which case the data will be pulled from the INPE website for the corresponding biome.

It can also be a path to a .zip file as is obtained from the INPE website, containing a file named "deter_public.dbf".}

\item{years}{A numeric \code{vector} of years, used to filter the data.}
}
\value{
A \code{tibble}
}
\description{
Loads INPE deforestation data.
}
\examples{
load_deter_raw("amazonia")

load_deter_raw("cerrado", c(2016,2017,2018))

load_deter_raw("C:\\\\...\\\\'deter'.zip")

}
\seealso{
\link{load_deter}, for loading and treating the data.
}
\author{
DataZoom, Department of Economics, Pontifical Catholic University of Rio de Janeiro.
}
