% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_ee.R
\name{load_earth_engine}
\alias{load_earth_engine}
\title{VIIRS Day/Night Band}
\usage{
load_earth_engine(
  db = "NOAA/VIIRS/DNB/MONTHLY_V1/VCMSLCFG",
  var = "avg_rad",
  year_begin = "2014",
  year_end = "2018",
  type = "municipality",
  state = "all",
  nclusters = 0
)
}
\arguments{
\item{db}{Google Earth Engine Data to be imported (TO-DO: Include support for other versions)}

\item{var}{Variable belonging to the db dataset}

\item{year_begin}{Year Begin}

\item{year_end}{Year End}

\item{type}{Geometry Type. Can be either "municip" or "uf". (TO-DO: Next version will include support for rasters)}

\item{state}{Municipalities of Interest; Can be 'all', 'RJ', ... (TO-DO: Include support for Legal Amazon)}

\item{nclusters}{Numbers of cluster to be used (TO:DO: Still need to implement this properly)}
}
\value{
A \code{tibble}
}
\description{
Monthly average radiance composite images using nighttime data from the
Visible Infrared Imaging Radiometer Suite (VIIRS) Day/Night Band (DNB).
}
\examples{

\dontrun{
load_ee(db = 'NOAA/VIIRS/DNB/MONTHLY_V1/VCMSLCFG',state=c('RJ'))
}

}
\author{
DataZoom, Department of Economics, Pontifical Catholic University of Rio de Janeiro.
}
