% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{load_ee}
\alias{load_ee}
\title{VIIRS Day/Night Band}
\usage{
load_ee(
  db = "NOAA/VIIRS/DNB/MONTHLY_V1/VCMSLCFG",
  var = "avg_rad",
  year_begin = "2013",
  year_end = "2018",
  type = "munic"
)
}
\arguments{
\item{db}{Google Earth Engine Data to be imported}

\item{var}{Variable belonging to the db dataset}

\item{year_begin}{Year Begin}

\item{year_end}{Year End}

\item{type}{Geometry Type}
}
\value{

}
\description{
Monthly average radiance composite images using nighttime data from the
Visible Infrared Imaging Radiometer Suite (VIIRS) Day/Night Band (DNB).
}
\examples{

}
