% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_ibge_data.R
\name{load_employment}
\alias{load_employment}
\title{Gets formal employment data for Brazil}
\usage{
load_employment(
  years,
  sectors = FALSE,
  space_aggregation = "municipality",
  language = "eng"
)
}
\arguments{
\item{years}{A numeric vector with years of interest. Supported years 2006-2018.}

\item{sectors}{If sectors is set to TRUE, loads data separated by economics sectors.}

\item{space_aggregation}{A string that indicates the level of aggregation of the data. It can be by "Municipality" or
"State"}

\item{language}{A string that indicates in which language the data will be returned. The default is "eng", so your data will be returned in English.
The other option is "pt" for Portuguese.}
}
\value{
A \code{tibble} with (total) Salary (in R$1.000,00), (number of) Employed people and (number of) Firms.
}
\description{
CEMPRE: Cadastro Central de Empresas \url{https://sidra.ibge.gov.br/pesquisa/cempre/quadros/brasil/2018}
Tabela 6449 - Empresas e outras organizações, pessoal ocupado total, pessoal ocupado assalariado, salários e outras remunerações,
por seção, divisão, grupo e classe da classificação de atividades (CNAE 2.0)
}
\examples{
\dontrun{
load_employment(2017)
}
}
