% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipios.R
\name{load_gdp}
\alias{load_gdp}
\title{Gets GDP at current prices and Population data for Brazil}
\usage{
load_gdp(years, aggregation_level = "municipality", language = "eng")
}
\arguments{
\item{years}{A numeric vector with years of interest. Supported years are 2002-2017.}

\item{aggregation_level}{A string that indicates the level of aggregation of the data. It can be by "Municipality" or
"State"}

\item{language}{A string that indicates in which language the data will be returned. The default is "eng", so your data will be returned in English.
The other option is "pt" for Portuguese.}
}
\value{
A \code{tibble} with GDP (in thousands), Population and GDP Per Capita (in units).
}
\description{
Gets GDP at current prices and Population data for Brazil
}
\examples{
load_gdp(2017)
}
