% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibama_multas.R
\name{load_ibama_multas}
\alias{load_ibama_multas}
\title{IBAMA - Brazilian Institute for the Environment and Renewable Natural Resources}
\usage{
load_ibama_multas(
  dataset = NULL,
  raw_data = FALSE,
  geo_level = "municipality",
  uf = NULL,
  language = "pt"
)
}
\arguments{
\item{dataset}{A dataset name ("distributed_fines", "collected_fines")}

\item{raw_data}{A \code{boolean} setting the return of raw or processed data}

\item{geo_level}{A \code{string} that defines the geographic level of the data. Can be only "municipality".}

\item{uf}{A\code{string} that defines a specific state within Brazil. Can be only written with each state two-letter specification.}

\item{language}{A \code{string} that indicates in which language the data will be returned. Currently, only Portuguese ("pt") and English ("eng") are supported. Defaults for Portuguese.}
}
\value{
A \code{tibble} with the selected data.
}
\description{
Loads information on environmental distributed or collected fines at the municipality or state levels considering the Amazon region
}
\examples{
\dontrun{
# download distributed raw data from state Bahia
raw_ibama_multas <- load_ibama_multas(
  dataset = "distributed_fines",
  raw_data = TRUE,
  uf = "BA"
)
}

}
