% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imazon.R
\encoding{UTF-8}
\name{load_imazon}
\alias{load_imazon}
\title{IMAZON}
\usage{
load_imazon(
  dataset = "imazon_shp",
  raw_data = TRUE,
  geo_level = "municipality",
  language = "pt_br"
)
}
\arguments{
\item{dataset}{There is one dataset available ("imazon_shp")}

\item{raw_data}{A \code{boolean} setting the return of raw or processed data}

\item{geo_level}{A \code{string} that defines the geographic level of the data. Only "municipality" available.}

\item{language}{A \code{string} that indicates in which language the data will be returned. Currently, only Portuguese is supported.}
}
\value{
A \code{tibble} with the selected data.
}
\description{
Loads information on ...  See \url{http://www.ipsamazonia.org.br/}
}
\examples{
\dontrun{
# download raw data from 2014
imazon <- load_imazon(dataset = "imazon_shp")
}
}
