% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPS.R
\encoding{UTF-8}
\name{load_IPS_data}
\alias{load_IPS_data}
\title{Compiles information on the indicators part of  the Social Progress Index (SPI) for
the Amazon region}
\usage{
load_IPS_data(download_directory, language = "english")
}
\arguments{
\item{download_directory}{Directory where raw data should be stored.}

\item{language}{Language used in returned dataset. Use \code{language = "portuguese"} or
\code{language = "english"}. Defaults to English.}
}
\value{
A data frame with information on the 60+ indicators part of the Social
Progress Index (SPI) for 2014 and 2018 considering municipalities part of the Legal
Amazon Region
}
\description{
The original data and more information on the construction of the indicator
can be found \href{http://www.ipsamazonia.org.br/}{here}
}
\examples{

load_IPS_data(
  download_directory = getwd(),
  language = "portuguese"
)

}
