% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbiomas.R
\encoding{UTF-8}
\name{load_mapbiomas_transition}
\alias{load_mapbiomas_transition}
\title{load_mapbiomas_transition}
\usage{
load_mapbiomas_transition(
  space_aggregation = c("municipality", "state", "municipio", "estado"),
  path = NULL,
  transition_interval = NULL
)
}
\arguments{
\item{space_aggregation}{A string that indicates the level of aggregation of the data. It can be by municipality or state}

\item{path}{A string indicating where the raw data is in your computer. The default is NULL which means the data will be extracted directly from the website}

\item{transition_interval}{A numeric object containing the desired interval in years to observe transitions desired in the data base (1, 2, 5 or 10), or have all included (NULL)}
}
\value{
A data base with data containing the area of each selected type of soil covering transition in each selected year
}
\description{
Download and filter data on transition of types of soil covering by year
}
\examples{
load_mapbiomas_transition(space_aggregation = 'state', path = NULL, transition_interval = 5)
}
\author{
DataZoom, Department of Economics, Pontifical Catholic University of Rio de Janeiro
}
