% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipios.R
\name{load_pib}
\alias{load_pib}
\title{Gets GDP at current prices for the Brazilian Legal Amazon}
\usage{
load_pib(years, aggregation_level = "City", language = "eng")
}
\arguments{
\item{years}{A numeric vector with years of interest. Supported years are 2002-2017.}

\item{aggregation_level}{A string that indicates the level of aggregation of the data. It can be by "City" or
"State"}

\item{language}{A string that indicates in which language the data will be returned. The default is "eng", so your data will be returned in English.
The other option is "pt" for Portuguese.}
}
\value{
A \code{tibble}.
}
\description{
Gets GDP at current prices for the Brazilian Legal Amazon
}
\examples{
load_pib(2017)
}
