% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodes.R
\encoding{UTF-8}
\name{load_prodes_raw}
\alias{load_prodes_raw}
\title{Loads deforestation data from INPE.}
\usage{
load_prodes_raw(source)
}
\arguments{
\item{source}{A number of different sources are supported:

Passing a numeric \code{vector} of years will download the corresponding data from the INPE website (data is organized by prodes-year - from august to july - for more information, see the vignette).

Passing a \code{string} with a directory's path will read data from all files named
"\{path\}/DesmatamentoMunicipiosXXXX.txt".

Alternatively, \code{source} may be a list of full file paths, or anything else readable by \code{utils::read.csv()}.}
}
\value{
A list of \code{tibble}.
}
\description{
Loads deforestation data from INPE.
}
\examples{
\dontrun{load_prodes_raw(2017)}

\dontrun{load_prodes_raw(c(2016, 2017, 2018))}

\dontrun{load_prodes_raw(system.file("extdata", package = "datazoom.amazonia"))}

\dontrun{load_prodes_raw(
  system.file("extdata", "DesmatamentoMunicipios2015.txt", package = "datazoom.amazonia")
)}
}
\seealso{
\link{load_prodes} for loading and treating the data.
}
\author{
DataZoom, Department of Economics, Pontifical Catholic University of Rio de Janeiro.
}
