% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siga.R
\encoding{UTF-8}
\name{load_siga}
\alias{load_siga}
\title{SIGA}
\usage{
load_siga(dataset, raw_data = FALSE, time_period = NA, language = "pt")
}
\arguments{
\item{dataset}{A dataset name ("HS92").}

\item{raw_data}{A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.}

\item{time_period}{A \code{numeric} indicating for which years the data will be loaded, in the format YYYY. Can be any vector of numbers, such as 2010:2012.}

\item{language}{A \code{string} that indicates in which language the data will be returned. Portuguese ("pt") and English ("eng") are supported.}
}
\description{
ANEEL's Generation Information System
}
\examples{
\dontrun{
# download treated data
clean_siga <- load_siga(
  dataset = "SIGA",
  raw_data = FALSE,
)
}

}
