% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipalities_biomes}
\alias{municipalities_biomes}
\title{IBGE codes and MAPBIOMAS id of Brazilian municipalities and biomes}
\format{
A data frame with 6537 rows and 4 variables:
\describe{
\item{feature_id}{MAPBIOMAS biome-municipality ID}
\item{code_muni}{IBGE 7-digit municipality code}
\item{abbrev_state}{state abbreviations (e.g. "AM")}
\item{municipality_mapbiomas}{municipality name in MAPBIOMAS data}
\item{biome}{biome}
}
}
\source{
Package \code{geobr} and \url{https://mapbiomas.org/}
}
\usage{
municipalities_biomes
}
\description{
A dataset containing each municipality-biome's IBGE code, state, biome, name and MAPBIOMAS ID. Mostly for our functions' internal use.
}
\keyword{datasets}
