% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipios.R
\name{load_amazon_gdp}
\alias{load_amazon_gdp}
\title{Gets GDP at current prices and Population data for the Brazilian Legal Amazon}
\usage{
load_amazon_gdp(years, space_aggregation = "municipality", language = "eng")
}
\arguments{
\item{years}{A numeric vector with years of interest. Supported years 2002-2017 for GDP data and 2001-2009, 2011-present for population data.}

\item{space_aggregation}{A string that indicates the level of aggregation of the data. It can be by "Municipality" or
"State"}

\item{language}{A string that indicates in which language the data will be returned. The default is "eng", so your data will be returned in English.
The other option is "pt" for Portuguese.}
}
\value{
A \code{tibble} with GDP (in thousands), Population and GDP Per Capita (in units).
}
\description{
Gets GDP at current prices and Population data for the Brazilian Legal Amazon
}
\examples{
\dontrun{load_amazon_gdp(2017)}
}
