% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aneel.R
\name{load_aneel}
\alias{load_aneel}
\title{ANEEL}
\usage{
load_aneel(dataset, raw_data = FALSE, language = "eng", year = NULL)
}
\arguments{
\item{dataset}{A dataset name ("energy_development_budget", "energy_generation" or "energy_enterprises_distributed")}

\item{raw_data}{A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.}

\item{language}{A \code{string} that indicates in which language the data will be returned. Portuguese ("pt") and English ("eng") are supported.}

\item{year}{A numeric value or vector of years (2017-2024).
Required for the "energy_development_budget" dataset.
Ignored for the other datasets.}
}
\value{
A \code{data.frame}: the raw source data when \code{raw_data = TRUE}, or a treated \code{tibble} when \code{raw_data = FALSE}.
}
\description{
National Electric Energy Agency - ANEEL
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
### DO NOT RUN ###
# download treated data about energy generation
clean_aneel <- load_aneel(
  dataset = "energy_generation",
  raw_data = FALSE
)

# download raw annual CDE budget data
raw_cde <- load_aneel(
  dataset = "energy_development_budget",
  year = 2021,
  raw_data = TRUE
)
\dontshow{\}) # examplesIf}
}
