% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aneel.R
\name{load_aneel}
\alias{load_aneel}
\title{ANEEL}
\usage{
load_aneel(dataset, raw_data = FALSE, language = "eng")
}
\arguments{
\item{dataset}{A dataset name ("energy_development_budget", "energy_generation" or "energy_enterprises_distributed")}

\item{raw_data}{A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.}

\item{language}{A \code{string} that indicates in which language the data will be returned. Portuguese ("pt") and English ("eng") are supported.}
}
\description{
National Electric Energy Agency - ANEEL
}
\examples{
\dontrun{
# download treated data about energy generation
clean_aneel <- load_aneel(
  dataset = "energy generation",
  raw_data = FALSE
)
}

}
