% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ben.R
\name{load_ben}
\alias{load_ben}
\title{BEN}
\usage{
load_ben(
  dataset = "ben",
  raw_data = FALSE,
  time_period = 2004:2021,
  language = "pt"
)
}
\arguments{
\item{dataset}{A dataset name ("ben")}

\item{raw_data}{A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.}

\item{time_period}{A \code{numeric} indicating for which years the data will be loaded, in the format YYYY. Can be any vector of numbers, such as 2010:2012.}

\item{language}{Only available in Portuguese ("pt") as of now}
}
\description{
National Energy Balance
}
\examples{
\dontrun{
# download treated data for 2016
clean_ben <- load_ben(
  dataset = "ben",
  raw_data = FALSE,
  time_period = 2016
)
}

}
