% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comex.R
\encoding{UTF-8}
\name{load_br_trade}
\alias{load_br_trade}
\title{Download and treat Comex data}
\usage{
load_br_trade(
  dataset = NULL,
  raw_data = NULL,
  geo_level = NULL,
  time_period = 2018:2019,
  language = "eng",
  time_id = "year",
  prod_class = "hs2"
)
}
\arguments{
\item{dataset}{A string indicating the type of data to be downloaded, chosen between "comex_export_mun" (\strong{?}), "comex_import_mun" (\strong{?}), "comex_export_prod" (\strong{?}), "comex_import_prod" (\strong{?}).}

\item{raw_data}{A logical value indicating whether the data should be raw (\code{TRUE}) or treated (\code{FALSE}).}

\item{geo_level}{\strong{NOT USED IN THIS FUNCTION}.}

\item{time_period}{A \code{vector} of type \code{numeric} indicating the desired years to be downloaded. Defaults to 2018 and 2019 (\strong{WHY?}).}

\item{language}{A string indicating the language to be used in variable names, chosen between "en" (english) and "pt" (portuguese). Defaults to "en".}

\item{time_id}{\strong{NOT USED IN THIS FUNCTION}.}

\item{prod_class}{A string indicating the classification to be downloaded, chosen between "hs" (SH - Sistema Harmonizado), "cuci" (CUCI - Classificação Uniforme do Comércio Internacional), "isic" (ISIC - Classificação Internacional Padrão por Atividade Econômica), "cgce" (CGCE - Classificação por Grandes Categorias Econômicas). Defaults to "hs2" (\strong{WHY?}).}
}
\value{
A tibble with the data.
}
\description{
Download and treat both imports and exports data from Comex.
}
\examples{
# download treated exports data by municipality from 1997 to 2021
exp_mun <- load_br_trade(dataset = "comex_export_mun", raw_data = FALSE, time_period = 1997:2021)

# download raw imports data by municipality from 1997 to 2021
raw_imp_mun <- load_br_trade(dataset = "comex_import_mun", raw_data = TRUE, time_period = 1997:2021)

# download treated imports data by municipality from 1997 to 2021 using "CUCI" classification
imp_mun_cuci <- load_br_trade(dataset = "comex_import_mun", raw_data = FALSE, time_period = 1997:2021, prod_class = "cuci")

}
\seealso{
\url{https://www.gov.br/produtividade-e-comercio-exterior/pt-br/assuntos/comercio-exterior/estatisticas/base-de-dados-bruta}
}
