% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censoagro.R
\name{load_censoagro}
\alias{load_censoagro}
\title{Censo Agropecuario}
\usage{
load_censoagro(
  dataset,
  raw_data = FALSE,
  geo_level,
  time_period,
  language = "eng"
)
}
\arguments{
\item{dataset}{A dataset name ("HS92").}

\item{raw_data}{A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.}

\item{geo_level}{A \code{string} that defines the geographic level of the data. Can be one of "country", "state" , "region" or "municipality".}

\item{time_period}{A \code{numeric} indicating for which years the data will be loaded, in the format YYYY. Can be any vector of numbers, such as 2010:2012.}

\item{language}{A \code{string} that indicates in which language the data will be returned. Portuguese ("pt") and English ("eng") are supported.}
}
\value{
A \code{tibble}.
}
\description{
Loads information on agricultural establishments and activities

#' @param dataset A dataset name ("land_area_total", "area_use", "employess_tractors","land_area_producer_condition","animal_specie_production",
"vegetable_production_area_type", "land_area_total_mean", "use_type", "employess_total_mean", "tractors_total_mean",
"bovine_area_mean", "animal_herd_type", "income_mean_vegetable_type", "vegetable_area_income_coffee_orange",
"production_permanent_crops", "production_temporary_crops", "livestock_production").
}
\examples{
\dontrun{
# Download total land area data at the country level in year 2006
data <- load_censoagro(
  dataset = "land_area_total",
  raw_data = TRUE,
  geo_level = "country",
  time_period = 2006
)

# Download temporary production crops data by state (geo_level = "state") in year 2006
  in portuguese (language = "pt").
data <- load_censoagro(
  dataset = "production_temporary_crops",
  raw_data = FALSE,
  geo_level = "state",
  time_period = 1996,
  language = "pt"
)
}


## We should include support for microregion/mesoregion

## In case of dataset = "livestock_production", use the googledrive package to download data
}
