% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censoagro.R
\name{load_censoagro}
\alias{load_censoagro}
\title{Censo Agropecuario}
\usage{
load_censoagro(
  dataset = "censoagro",
  raw_data = FALSE,
  geo_level,
  time_period,
  language = "eng"
)
}
\arguments{
\item{dataset}{A dataset name ("censoagro").}

\item{raw_data}{A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.}

\item{geo_level}{A \code{string} that defines the geographic level of the data. Can be one of "country", "state" or "municipality".}

\item{time_period}{A \code{numeric} indicating for which years the data will be loaded, in the format YYYY. Can be any vector of numbers, such as 2010:2012.}

\item{language}{A \code{string} that indicates in which language the data will be returned. Portuguese ("pt") and English ("eng") are supported.}
}
\value{
A \code{tibble}.
}
\description{
Loads information on
}
\examples{
\dontrun{
# Download total land area data at the country level in year 2006
data <- load_censoagro(
  dataset = "land_area_total",
  raw_data = TRUE,
  geo_level = "country",
  time_period = 2006
)

# Download temporary production crops data by state (geo_level = "state") in year 2006
  in portuguese (language = "pt").
data <- load_censoagro(
  dataset = "production_temporary_crops",
  raw_data = FALSE,
  geo_level = "state",
  time_period = 2008:2010,
  language = "pt"
)
}

}
