% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasus.R
\name{load_datasus}
\alias{load_datasus}
\title{DATASUS - No longer available}
\usage{
load_datasus(
  dataset = NULL,
  raw_data = NULL,
  time_period = NULL,
  language = "eng"
)
}
\arguments{
\item{dataset}{Ignored. Kept only for compatibility.}

\item{raw_data}{Ignored. Kept only for compatibility.}

\item{time_period}{Ignored. Kept only for compatibility.}

\item{language}{Ignored. Kept only for compatibility.}
}
\value{
\code{NULL}. Always returns empty.
}
\description{
The \code{load_datasus} function is no longer part of the package.
Updates on the matter coming soon.
It is kept only for compatibility and will return a warning when called.
}
