% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrad.R
\encoding{UTF-8}
\name{load_degrad}
\alias{load_degrad}
\title{Loads and cleans degradation data from INPE.}
\usage{
load_degrad(source, aggregation_level = "municipality", language = "eng")
}
\arguments{
\item{source}{A number of different sources are supported:

Passing a numeric \code{vector} of years will download the corresponding data from the INPE website. Available years are 2007-2016.

Passing a \code{string} with a directory's path will read data from all shapefiles in the directory.

Alternatively, \code{source} may be a list of full file paths, or anything else readable by \code{sf::read_sf()}.}

\item{aggregation_level}{A string that indicates the level of aggregation of the data. It can be by "Municipality" or
"State".}

\item{language}{A string that indicates in which language the data will be returned. The default is "eng", so your data will be returned in English.
The other option is "pt" for Portuguese.}
}
\value{
A \code{tibble}.
}
\description{
Loads and cleans degradation data from INPE.
}
\examples{
load_degrad(2016)

load_degrad(
  c(2015, 2016),
  aggregation_level = "state",
  language = "pt"
)

\dontrun{
load_degrad(
  "~/Downloads",
  aggregation_level = "state",
  language = "en"
)

load_degrad(
  "~/Downloads/degrad2016_final_shp/DEGRAD_2016_pol.shp",
  aggregation_level = "municipality",
  language = "pt"
)
}
}
\seealso{
\link{load_degrad_raw} for loading raw data.
}
\author{
DataZoom, Department of Economics, Pontifical Catholic University of Rio de Janeiro.
}
