% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deter.R
\encoding{UTF-8}
\name{load_deter}
\alias{load_deter}
\title{DETER - Forest Degradation in the Brazilian Amazon

Loads information on change in forest cover in the Amazon. Survey is done at state or municipal level. See \url{http://www.obt.inpe.br/OBT/assuntos/programas/amazonia/deter/deter}}
\usage{
load_deter(
  dataset = NULL,
  raw_data = NULL,
  geo_level = "municipality",
  time_period = "all",
  language = "eng",
  time_id = "year"
)
}
\arguments{
\item{dataset}{A dataset name (\code{deter_amz}, \code{deter_cerrado} with information about both Amazon and Cerrado}

\item{raw_data}{A \code{boolean} setting the return of raw or processed data}

\item{geo_level}{A \code{string} that defines the geographic level of the data. Can be municipality or state.}

\item{time_period}{A \code{numeric} indicating what years will the data be loaded in the format YYYY. Default is all time-periods}

\item{language}{A \code{string} that indicates in which language the data will be returned. Currently, only Portuguese and English are supported.}

\item{time_id}{A \code{numeric} that defines whether data will be aggregated by month or year.}
}
\value{
A \code{tibble} with a panel of N x T observations
}
\description{
DETER - Forest Degradation in the Brazilian Amazon

Loads information on change in forest cover in the Amazon. Survey is done at state or municipal level. See \url{http://www.obt.inpe.br/OBT/assuntos/programas/amazonia/deter/deter}
}
