% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deter.R
\encoding{UTF-8}
\name{load_deter}
\alias{load_deter}
\title{Loads and cleans INPE data on areas with deforestation warnings}
\usage{
load_deter(
  source = "amazonia",
  aggregation_level = "municipality",
  time_aggregation = "year",
  language = "eng"
)
}
\arguments{
\item{source}{A string indicating where the data will be drawn from.

It can be "Amazonia" or "Cerrado", in which case the data will be pulled from the INPE website for the corresponding biome. Set to "Amazonia" by default.

It can also be a path to a .zip file as is obtained from the INPE website, containing a file named "deter_public.dbf".}

\item{aggregation_level}{A string that indicates the level of aggregation of the data. It can be by "Municipality", the default alternative, or
also "State".}

\item{time_aggregation}{A string defining whether data will be aggregated by "month" or by "year".}

\item{language}{A string that indicates in which language the data will be returned. The default is "eng", so your data will be returned in English.
The other option is "pt" for Portuguese.}
}
\value{
A \code{tibble}.
}
\description{
Loads and cleans INPE data on areas with deforestation warnings
}
\examples{
load_deter()
\dontrun{
load_deter("path/to/deter.zip", aggregation_level = "municipality", time_aggregation = "year")
}

load_deter(
  source = "cerrado",
  aggregation_level = "state",
  time_aggregation = "month",
  language = "pt"
)
}
\seealso{
\link{load_deter_raw} for loading raw data.
}
\author{
DataZoom, Department of Economics, Pontifical Catholic University of Rio de Janeiro.
}
