% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deter.R
\encoding{UTF-8}
\name{load_deter}
\alias{load_deter}
\title{Loads and cleans deforestation data from INPE.}
\usage{
load_deter(
  source,
  aggregation_level = "municipality",
  years = "all",
  language = "eng"
)
}
\arguments{
\item{source}{A string indicating where the data will be drawn from.

It can be "Amazonia" or "Cerrado", in which case the data will be pulled from the INPE website for the corresponding biome.

It can also be a path to a .zip file as is obtained from the INPE website, containing a file named "deter_public.dbf".}

\item{aggregation_level}{A string that indicates the level of aggregation of the data. It can be by "Municipality" or
"State".}

\item{years}{A numeric \code{vector} of years, used to filter the data.}

\item{language}{A string that indicates in which language the data will be returned. The default is "eng", so your data will be returned in English.
The other option is "pt" for Portuguese.}
}
\value{
A \code{tibble}.
}
\description{
Loads and cleans deforestation data from INPE.
}
\examples{
load_deter("amazonia")

load_deter("cerrado", years = 2018)

load_deter("C:\\\\...\\\\'deter'.zip") 

load_deter(
source = "cerrado",
aggregation_level = "state",
years = c(2017,2018),
language = "pt"
) 

}
\seealso{
\link{load_deter_raw} for loading raw data.
}
\author{
DataZoom, Department of Economics, Pontifical Catholic University of Rio de Janeiro.
}
