% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe.R
\encoding{UTF-8}
\name{load_epe}
\alias{load_epe}
\title{EPE}
\usage{
load_epe(
  dataset,
  raw_data = FALSE,
  time_period = "default",
  geo_level = "State",
  language = "pt"
)
}
\arguments{
\item{dataset}{A dataset name, ("energy_consumption_per_class") or ("national_energy_balance")}

\item{raw_data}{A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.}

\item{time_period}{A \code{numeric} indicating for which years the data will be loaded, in the format YYYY. Can be any vector of numbers, such as 2010:2012.}

\item{geo_level}{A geographical level, ("State") or ("Subsystem"), only available for ("energy_consumption_per_class")}

\item{language}{Only available in Portuguese ("pt") as of now}
}
\description{
Electrical Energy Monthly Consumption per Class
}
\examples{
\dontrun{
# download treated data for 2016
clean_epe <- load_epe(
  dataset = "energy_consumption_per_class",
  geo_level = "State",
  raw_data = FALSE,
  time_period = 2016
)
}

}
