% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe.R
\name{load_epe}
\alias{load_epe}
\title{EPE}
\usage{
load_epe(dataset, geo_level = "state", raw_data = FALSE, language = "eng")
}
\arguments{
\item{dataset}{Dataset name: "consumer_energy_consumption", "industrial_energy_consumption", or "national_energy_balance"}

\item{geo_level}{Geographical level: "state" or "subsystem". Only applies to consumer or industrial datasets.}

\item{raw_data}{If TRUE, returns raw sheets.}

\item{language}{"pt" or "en". Applies to processed outputs.}
}
\description{
Electrical Energy Monthly Consumption per Class or Industrial Sector
}
