% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe.R
\encoding{UTF-8}
\name{load_epe}
\alias{load_epe}
\title{EPE}
\usage{
load_epe(dataset, raw_data = FALSE, time_period = 2004:2021, language = "pt")
}
\arguments{
\item{dataset}{A dataset name, either ("CONSUMO") for consumption, or ("CONSUMIDOR") for consumers}

\item{raw_data}{A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.}

\item{time_period}{A \code{numeric} indicating for which years the data will be loaded, in the format YYYY. Can be any vector of numbers, such as 2010:2012.}

\item{language}{Only available in Portuguese ("pt") as of now}
}
\description{
Electrical Energy Monthly Consumption per Class
}
\examples{
\dontrun{
# download treated data for 2016
clean_epe <- load_epe(
  dataset = "CONSUMO",
  raw_data = FALSE,
  time_period = 2016
)
}

}
