% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe.R
\name{load_epe}
\alias{load_epe}
\title{EPE}
\usage{
load_epe(dataset, geo_level = "state", raw_data = FALSE, language = "eng")
}
\arguments{
\item{dataset}{Dataset name: "consumer_energy_consumption", "industrial_energy_consumption", or "national_energy_balance"}

\item{geo_level}{Geographical level: "state" or "subsystem". Only applies to consumer or industrial datasets.}

\item{raw_data}{A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.}

\item{language}{A \code{string} that indicates in which language the data will be returned. Portuguese ("pt") and English ("eng") are supported.}
}
\description{
Electrical Energy Monthly Consumption per Class or Industrial Sector
}
\examples{
\dontrun{
# download treated (raw_data = FALSE) data about
# consumer energy consumption (dataset = "consumer_energy_consumption")
# at the state level (geo_level = "state")
data <- load_epe(
  dataset = "consumer_energy_consumption",
  geo_level = "state",
  raw_data = FALSE
)
# download treated (raw_data = FALSE) data
# from the National Energy Balance (dataset = "national_energy_balance")
balance <- load_epe(
  dataset = "national_energy_balance",
  raw_data = FALSE
)
}

}
