% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibama.R
\encoding{UTF-8}
\name{load_ibama}
\alias{load_ibama}
\title{Compiles environmental fines data in the Amazon Region from IBAMA}
\usage{
load_ibama(
  download_data = TRUE,
  download_directory,
  load_from_where,
  time_aggregation = c("year", "month"),
  space_aggregation = "municipality",
  years = 2005:2021,
  language = "eng"
)
}
\arguments{
\item{download_data}{If \code{TRUE}, downloads the data from the original source.
Else, loads data from \code{load_from_where} directory.}

\item{download_directory}{Directory where raw data should be stored.
Used in case \code{download_data = TRUE}.}

\item{load_from_where}{Original that should be loaded. Used in
case \code{download_data = FALSE}.}

\item{time_aggregation}{Temporal level of aggregation, choose \code{year} and/or \code{month}.
If \code{language = 'pt'}, pick "ano" and/or "mes'"}

\item{space_aggregation}{Geographical level of aggregation, choose \code{municipality} and/or
\code{State}. For Portuguese version, \code{municipio} and/or \code{uf}}

\item{years}{Time period to be considered. Contemplates 2005 to the present day}

\item{language}{Language used in returned dataset. Use \code{language = "pt"} or
\code{language = "eng"}}
}
\value{
A data frame with counts for variables related to environmental fines and
infractions by \code{time_aggregation}-\code{space_aggregation} chosen on the Amazon region.
Data include
counts for total number of infractions, infractions that already went to trial, and
number of unique perpetrators of infractions on the given place-time period.
}
\description{
Original data comes from \href{https://servicos.ibama.gov.br/ctf/publico/areasembargadas/ConsultaPublicaAreasEmbargadas.php}{IBAMA's website}
}
\examples{

load_ibama(
  download_directory = getwd(),
  time_aggregation = "year",
  space_aggregation = "state", years = 2005:2006
)

load_ibama(
  download_directory = getwd(),
  time_aggregation = "ano",
  space_aggregation = "municipio", years = c(2010, 2012),
  language = "pt"
)
\dontrun{

load_ibama(
  download_data = FALSE,
  load_from_where = "./Desktop/data.xls",
  time_aggregation = c("year", "month"),
  space_aggregation = c("municipality", "state")
)
}

}
