% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iema.R
\name{load_iema}
\alias{load_iema}
\title{IEMA - Institute of Environment and Water Resources}
\usage{
load_iema(dataset = "iema", raw_data = FALSE, language = "eng")
}
\arguments{
\item{dataset}{A dataset name ("iema")}

\item{raw_data}{A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.}

\item{language}{A \code{string} that indicates in which language the data will be returned. Portuguese ("pt") and English ("eng") are supported.}
}
\value{
A \code{tibble}.
}
\description{
Loads information on electric energy access at the municipality level in the Amazon region
}
\examples{
\dontrun{
# Download treated data
data <- load_iema(raw_data = FALSE)
}

}
