% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imazon.R
\name{load_imazon}
\alias{load_imazon}
\title{IMAZON - Deforestation pressure by municipality}
\usage{
load_imazon(dataset = "imazon_shp", raw_data = FALSE, language = "eng")
}
\arguments{
\item{dataset}{There is one dataset available ("imazon_shp")}

\item{raw_data}{A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.}

\item{language}{A \code{string} that indicates in which language the data will be returned. Portuguese ("pt") and English ("eng") are supported.}
}
\value{
A \code{tibble}.
}
\description{
Loads data categorizing each municipality by the level of deforestation pressure it faces
}
\examples{
\dontrun{
# Download treated data
data <- load_imazon(raw_data = FALSE)
}

}
