% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPS.R
\name{load_ips}
\alias{load_ips}
\title{IPS - Amazon Social Progress Index

Loads information on the social and environmental performance of the Legal Amazon. Survey is done at the municipal level and data is available in 2014 and 2018. See \url{http://www.ipsamazonia.org.br/}}
\usage{
load_ips(
  dataset = NULL,
  raw_data = NULL,
  geo_level = "municipality",
  time_period = c(2014, 2018),
  language = "eng"
)
}
\arguments{
\item{dataset}{A dataset name(\code{ips})

@param raw_data A \code{boolean} setting the return of raw or processed data

@param geo_level A \code{string} that defines the geographic level of the data. In this case is municipality

@param time_period A \code{numeric} indicating what years will the data be loaded in the format YYYY.

@param language A \code{string} that indicates in which language the data will be returned. Currently, only Portuguese and English are supported.

@encoding UTF-8

@return A \code{tibble} with a panel of N x T observations

@importFrom magrittr \%>\%

@export load_ips

@examples \dontrun{datazoom.amazonia::load_ips(dataset = 'ips', 'municipality', 2014, language = "pt")}}
}
\description{
IPS - Amazon Social Progress Index

Loads information on the social and environmental performance of the Legal Amazon. Survey is done at the municipal level and data is available in 2014 and 2018. See \url{http://www.ipsamazonia.org.br/}
}
