% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPS.R
\encoding{UTF-8}
\name{load_IPS}
\alias{load_IPS}
\title{Compiles information on the indicators part of  the Social Progress Index (SPI) for
the Amazon region}
\usage{
load_IPS(download_directory, language = "eng")
}
\arguments{
\item{download_directory}{Directory where raw data should be stored.}

\item{language}{Language used in returned dataset. Use \code{language = "pt"} or
\code{language = "eng"}. Defaults to English.}
}
\value{
A data frame with information on the 60+ indicators part of the Social
Progress Index (SPI) for 2014 and 2018 considering municipalities part of the Legal
Amazon Region. Columns indexed by "D" refer to each of the three dimensions of the index.
These dimenstions are split into twelve components, whose columns are indexed by "C".
The components, in turn, are composed of the raw indicators, whose columns are indexed
by "V"
}
\description{
The original data and more information on the construction of the indicator
can be found \href{http://www.ipsamazonia.org.br/}{here}
}
\examples{

\dontrun{
load_IPS(
  download_directory = "./Desktop",
  language = "pt"
)
}

}
