% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPS.R
\name{load_ips}
\alias{load_ips}
\title{IPS - Amazon Social Progress Index}
\usage{
load_ips(dataset = "all", raw_data = FALSE, time_period, language = "eng")
}
\arguments{
\item{dataset}{A dataset name ("all", "life_quality", "sanit_habit", "violence", "educ", "communic", "mortality", or "deforest")}

\item{raw_data}{A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.}

\item{time_period}{A \code{numeric} indicating for which years the data will be loaded, in the format YYYY. Can be any vector of numbers, such as 2010:2012.}

\item{language}{A \code{string} that indicates in which language the data will be returned. Portuguese ("pt") and English ("eng") are supported.}
}
\value{
A \code{tibble}.
}
\description{
Loads information on the social and environmental performance of the Legal Amazon.
}
\examples{
\dontrun{
# download raw data from 2014
ips <- load_ips(dataset = "all", raw_data = TRUE, time_period = 2014)
}

}
