% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\encoding{UTF-8}
\name{load_pam_temporary}
\alias{load_pam_temporary}
\title{PAM Temporary tillage

Loads agricultural data of temporary tillages from Brazil from 1974 to the present}
\usage{
load_pam_temporary(
  years,
  aggregation_level = "country",
  language = "pt",
  long = FALSE
)
}
\arguments{
\item{years}{A \code{vector} indicating what years will the data be loaded}

\item{aggregation_level}{A \code{string} that defines the geographic level of the data. Defaults to National level, but can be one of "country", "region", "state", "mesoregion", "microregion" and "city"}

\item{language}{A \code{string} that indicates in which language the data will be returned. The default is "pt", so your data will be returned in Portuguese. Currently, only Portuguese and English are supported.}

\item{long}{A \code{boolean} that sets the format of the returned data. \code{long = TRUE} will return a list where each data frame represents a variable and products are transformed to columns so each geographical aggregation level/year will only have a single line. Default value is \code{FALSE}.}
}
\value{
A \code{data frame} or a \code{list} of data frames if \code{long} is set to \code{TRUE}.
}
\description{
PAM Temporary tillage

Loads agricultural data of temporary tillages from Brazil from 1974 to the present
}
\examples{
\dontrun{
datazoom.amazonia::load_pam_temporary(2010, aggregation_level = "country")
}

}
\author{
DataZoom, Department of Economics, Pontifical Catholic University of Rio de Janeiro
}
