% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pevs.R
\encoding{UTF-8}
\name{load_pevs}
\alias{load_pevs}
\title{PEVS Plant extraction

Loads quantity produced and yield value of plant extraction, by type of product, from 1986 to 2019}
\usage{
load_pevs(
  type = NULL,
  geo_level = "municipality",
  time_period = 2018:2019,
  language = "pt"
)
}
\arguments{
\item{geo_level}{A \code{string} that defines the geographic level of the data. Defaults to National level, but can be one of "country", "region", "state", "mesoregion", "microregion" and "city"}

\item{time_period}{A \code{vector} indicating what years will the data be loaded}

\item{language}{A \code{string} that indicates in which language the data will be returned. The default is "pt", so your data will be returned in Portuguese. Currently, only Portuguese and English are supported.}
}
\value{
A \code{data frame} or a \code{list} of data frames if \code{long} is set to \code{TRUE}.
}
\description{
PEVS Plant extraction

Loads quantity produced and yield value of plant extraction, by type of product, from 1986 to 2019
}
\examples{
\dontrun{datazoom.amazonia::load_pevs_vegextr(2013, aggregation_level = "country")}
}
\author{
DataZoom, Department of Economics, Pontifical Catholic University of Rio de Janeiro
}
