% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodes.R
\name{load_prodes}
\alias{load_prodes}
\title{PRODES - Deforestation Monitoring Project in the Legal Amazon by Satellite}
\usage{
load_prodes(dataset = "deforestation", raw_data = FALSE, language = "eng")
}
\arguments{
\item{dataset}{A dataset name ("deforestation").}

\item{raw_data}{A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.}

\item{language}{A \code{string} that indicates in which language the data will be returned. Portuguese ("pt") and English ("eng") are supported.}
}
\value{
A \code{tibble} with the selected data.
}
\description{
Loads data on deforestation in the Legal Amazon region.
}
\examples{
\dontrun{
# Download treated data (raw_data = FALSE)
# in portuguese (language = 'pt').
data <- load_prodes(
  raw_data = FALSE,
  language = "pt"
)
}

}
