% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodes.R
\encoding{UTF-8}
\name{load_prodes}
\alias{load_prodes}
\title{PRODES -Deforestation Monitoring Project in the Legal Amazon by Satellite

Loads information on clearcut deforestation in the Legal Amazon and annual deforestation rates in the region. Survey is done at state or municipality level and data is available from 2000 to 2020. See \url{http://www.dpi.inpe.br/prodesdigital}}
\usage{
load_prodes(
  dataset = NULL,
  raw_data = NULL,
  geo_level = "municipality",
  time_period = 2010:2012,
  language = "pt"
)
}
\arguments{
\item{dataset}{A dataset name (\code{prodes_accum_deforestation}, \code{prodes_forest}, \code{prodes_hydrography}, \code{prodes_annual_increase_deforestation}, \code{prodes_cloud})}

\item{raw_data}{A \code{boolean} setting the return of raw or processed data}

\item{geo_level}{A \code{string} that defines the geographic level of the data.Can be "state" or "municipality"}

\item{time_period}{A \code{numeric} indicating what years will the data be loaded in the format YYYY. Can be a sequence of numbers such as 2010:2012.}

\item{language}{A \code{string} that indicates in which language the data will be returned. Currently, only Portuguese and English are supported.}
}
\value{
A \code{tibble} with a panel of N x T observations by municipality-year
}
\description{
PRODES -Deforestation Monitoring Project in the Legal Amazon by Satellite

Loads information on clearcut deforestation in the Legal Amazon and annual deforestation rates in the region. Survey is done at state or municipality level and data is available from 2000 to 2020. See \url{http://www.dpi.inpe.br/prodesdigital}
}
\examples{
\dontrun{
# download treated data from 2000 to 2020
prodes_all <- load_prodes(dataset = "prodes", raw_data = FALSE, time_period = 2000:2020)

# download raw data from 2000 to 2020
raw_prodes_all <- load_prodes(dataset = "prodes", raw_data = TRUE, time_period = 2000:2020)
}


}
