% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodes.R
\name{load_prodes}
\alias{load_prodes}
\title{PRODES - Deforestation Monitoring Project in the Legal Amazon by Satellite}
\usage{
load_prodes(
  dataset = "deforestation",
  raw_data = FALSE,
  time_period = 2023,
  language = "eng"
)
}
\arguments{
\item{dataset}{A dataset name. Can be one of "deforestation", "residual_deforestation", "native_vegetation", "hydrography", "non_forest", or "clouds".}

\item{raw_data}{A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.}

\item{time_period}{A \code{numeric} indicating for which years the data will be loaded, in the format YYYY. Can be any vector of numbers, such as 2010:2012.
\itemize{
\item Between 2007 - 2023 for dataset "deforestation". Deforestation for 2007 includes all cumulative deforestation up to 2007. For other years, deforestation is incremental
\item Between 2010 - 2023 for dataset "residual_deforestation"
\item Only 2023 for all other datasets
}}

\item{language}{A \code{string} that indicates in which language the data will be returned. Portuguese ("pt") and English ("eng") are supported.}
}
\value{
A \code{tibble} with the selected data if raw_data is \code{FALSE}, and a \code{SpatRaster} is \code{TRUE}.
}
\description{
Loads data on deforestation in the Legal Amazon region.
}
\examples{
\dontrun{
# Download treated data (raw_data = FALSE)
# in portuguese (language = 'pt').
data <- load_prodes(
  dataset = "deforestation",
  raw_data = FALSE,
  time_period = 2020:2023,
  language = "pt"
)
}

}
