% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeg.R
\encoding{UTF-8}
\name{load_seeg}
\alias{load_seeg}
\title{Greenhouse gases emission estimates (SEEG)

Loads data of estimates of emission of greenhouse gases of Brazilian cities}
\usage{
load_seeg(language = "pt")
}
\arguments{
\item{language}{A \code{string} that indicates in which language the data will be returned. The default is "pt", so your data will be returned in Portuguese. Currently, only Portuguese and English are supported.}
}
\value{
A \code{data frame}.
}
\description{
Greenhouse gases emission estimates (SEEG)

Loads data of estimates of emission of greenhouse gases of Brazilian cities
}
\examples{
\dontrun{datazoom.amazonia::load_seeg()}

}
\author{
DataZoom, Department of Economics, Pontifical Catholic University of Rio de Janeiro
}
