% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeg.R
\encoding{UTF-8}
\name{load_seeg}
\alias{load_seeg}
\title{Greenhouse gases emission estimates (SEEG)}
\usage{
load_seeg(
  dataset = "seeg",
  raw_data = FALSE,
  geo_level = "municipality",
  language = "eng"
)
}
\description{
Loads data of estimates of emission of greenhouse gases of Brazilian cities
}
\examples{
\dontrun{datazoom.amazonia::load_seeg()}

}
\author{
DataZoom, Department of Economics, Pontifical Catholic University of Rio de Janeiro
}
