% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siga.R
\name{load_siga}
\alias{load_siga}
\title{SIGA}
\usage{
load_siga(dataset = "siga", raw_data = FALSE, language = "pt")
}
\arguments{
\item{dataset}{A dataset name ("siga")}

\item{raw_data}{A \code{boolean} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) data.}

\item{language}{Only available in Portuguese ("pt") as of now}
}
\description{
ANEEL's Generation Information System
}
\examples{
\dontrun{
# download treated data
clean_siga <- load_siga()
}

}
